Updated the parser to include detailed error checking, and correctly compile information about each activity. Only thing that remains it to link it up to the other classes.

+++ HealthTrackerApp/src/main/java/seng202/group8/parser/Distance.java
@@ -2,37 +2,37 @@ package seng202.group8.parser;
 
 public class Distance {
     // hello
-    private int latitude;
-    private int longitude;
-    private int altitude;
+    private double latitude;
+    private double longitude;
+    private double altitude;
 
-    public Distance(int temp_lat, int temp_long, int temp_alt) {
+    public Distance(double temp_lat, double temp_long, double temp_alt) {
         latitude = temp_lat;
         longitude = temp_long;
         altitude = temp_alt;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public int getAltitude() {
+    public double getAltitude() {
         return altitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public void setAltitude(int altitude) {
+    public void setAltitude(double altitude) {
         this.altitude = altitude;
     }
 }
